+++ main-service/src/main/java/ru/practicum/mainservice/model/request/UpdateEventUserRequest.java
@@ -10,7 +10,7 @@ import lombok.experimental.FieldDefaults;
 @NoArgsConstructor
 public class UpdateEventUserRequest {
     String annotation;
-    Long category;
+    Integer category;
     String description;
     String eventDate;
     LocationDto location;
